refactored edit ingredients component
