+++ src/components/create-recipe/edit-ingredients.js
@@ -1,11 +1,13 @@
 import React, {useState} from "react";
 import { createRecipeIngredient, deleteRecipeIngredient } from "../../reducers/recipe-ingredients-reducer";
 import {useDispatch, useSelector} from "react-redux";
-import { createRecipeIngredientThunk, deleteRecipeIngredientThunk, findIngredientsForRecipeThunk } from "../recipe-ingredients-thunk";
+import { createRecipeIngredientThunk, deleteRecipeIngredientThunk, findIngredientsForRecipeThunk } from "../../services/recipe-ingredients-thunk";
 import "./index.css";
 
 const EditIngredients = ({recipe}) => {
     const ingredients = useSelector(state => state.recipeIngredients);
+    console.log(recipe);
+    console.log(ingredients);
     const [editIngredient, setIngredient] = useState('');
     const dispatch = useDispatch();
 
@@ -35,7 +37,7 @@ const EditIngredients = ({recipe}) => {
             </div>
             <ul className="text-dark mt-2 p-0">
                 {
-                    ingredients.filter(ingredient => ingredient.rid === recipe._id)
+                    ingredients.recipeIngredients.filter(ingredient => ingredient.rid === recipe._id)
                     .map(ingredient => 
                     <li key={ingredient._id} className="list-group-item">
                         <div className="row pt-3">
